+++ R/annual-wages-by-wage-group.R
@@ -23,14 +23,16 @@
 #' @return \code{tbl_df} with data filtered by the selected criteria.
 #' @references \href{https://www.epi.org/data/}{Economic Policy Institute Data Library}
 #' @note Data source: SSA | Kopczuk, Saez, and Song (2010)
+#' @return data frame
 #' @export
 #' @examples
-#' get_annual_wages_by_wage_group()
+#' if (not_dos()) get_annual_wages_by_wage_group()
 get_annual_wages_by_wage_group <- function() {
 
   params <- list(subject="wagegroup")
 
   res <- epi_query(params)
+  if (is.null(res)) return(data.frame())
 
   cols <- stringi::stri_trans_tolower(res$columns$name)
   cols <- stringi::stri_replace_all_regex(cols, "[\\('%\\)]", "")
@@ -43,8 +45,7 @@ get_annual_wages_by_wage_group <- function() {
   out <- dplyr::mutate_all(out, "clean_cols")
   out <- suppressMessages(readr::type_convert(out))
 
-  cite <- html_text(read_html(res$meta$source %||% "<p>Economic Policy Institute</p>"))
-  message(sprintf('Note: %s\nCitation: "%s"', res$meta$notes %||% "None", cite))
+  show_citation(res)
 
   out