+++ R/compensation-wages-benefits.R
@@ -9,14 +9,16 @@
 #' @return \code{tbl_df} with data filtered by the selected criteria.
 #' @references \href{https://www.epi.org/data/}{Economic Policy Institute Data Library}
 #' @note Data source: NIPA | ECEC
+#' @return data frame
 #' @export
 #' @examples
-#' get_compensation_wages_and_benefits()
+#' if (not_dos()) get_compensation_wages_and_benefits()
 get_compensation_wages_and_benefits <- function() {
 
   params <- list(subject="compben")
 
   res <- epi_query(params)
+  if (is.null(res)) return(data.frame())
 
   cols <- stringi::stri_trans_tolower(res$columns$name)
   cols <- stringi::stri_replace_all_regex(cols, "[\\('%,\\)]", "")
@@ -32,8 +34,7 @@ get_compensation_wages_and_benefits <- function() {
   out <- dplyr::mutate_all(out, "clean_cols")
   out <- suppressMessages(readr::type_convert(out))
 
-  cite <- html_text(read_html(res$meta$source %||% "<p>Economic Policy Institute</p>"))
-  message(sprintf('Note: %s\nCitation: "%s"', res$meta$notes %||% "None", cite))
+  show_citation(res)
 
   out